Provide interface callbacks instead of implementing HAL callbacks
To provide more flexible use, provide an interface for the CAN interrupts in the same way the Timer interrupt interface is done. That is, expose an interface that the user's application plugs into, rather than implement the HAL callbacks directly. This allows all sorts of user preferences, like using different busses for different functions, logging interrupts, combining handlers, compiling stats or installing a mock.  On the other hand, it's harder to get started from scratch, because you have to implement the HAL callbacks yourself if you have not already. Documentation will help.

Some maintenance notes on the interrupt functions:
  - tx doesn't use hcan and rx needn't, because CANModule_local already holds a reference. Could make that consistent either way.
  - CO_STM32_FDCAN_Driver or CO_STM32_CAN_Driver are defined, appropriately, in "CO_driver_target.h". That is included via CANopenNode by way of "CO_app_STM32.h" -> "CANopen.h" -> "301/CO_driver.h" -> "CO_driver_target.h". If that link were to break, compilation may still silently succeed because `#else` is used instead of `#elifdef CO_STM32_CAN_Driver`. Making it fussier might be safer...
  - differences between FDCAN and CAN are slight. Could just do the following once, and then consolidate the multiple declarations. But the flexibility is useful and future proofing for now.

```
#ifdef CO_STM32_FDCAN_Driver
#define CO_CAN_HANDLE_TYPE FDCAN_HandleTypeDef*
#elifdef CO_STM32_CAN_Driver
#define CO_CAN_HANDLE_TYPE CAN_HandleTypeDef*
#endif
```